[TML] rename concept SimpleNetworkVar to SimpleNetwork
