--- TreeModelLib/BelowgroundCompetition/SimpleNetworkVar/__init__.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-@date: 2021-Today
-@author: marie-christin.wimmler@tu-dresden.de
-"""
-
-from .SimpleNetworkVar import SimpleNetworkVar
